requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
pypdf
python-docx
streamlit
langchain
langchain-community
langchain-text-splitters
pydantic
